+++ CHANGELOG.md
@@ -1,5 +1,5 @@
 # Unreleased
-- [...]
+- [timePicker] Fix display on small devices
 
 # v0.3.0 (08/06/2018)
 Breaking changes: