+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 # Unreleased
 - [timePicker] Fix display on small devices
+- Remove deprecated updateColor() method
 
 # v0.3.0 (08/06/2018)
 Breaking changes: