+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 # Unreleased
 - [timePicker] Fix display on small devices
+- Remove deprecated updateColor() method
 - [itemChoice] Fix right content display
+
 # v0.3.0 (08/06/2018)
 Breaking changes:
 - [Autocomplete] Render empty search with an autocomplete list