+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 # Unreleased
-- [timePicker] Fix display on small devices
+- [TimePicker] Fix display on small devices
+- [Itinerary] Component now using an array to display places (no `departure` and `arrival` props anymore)
 
 # v0.3.0 (08/06/2018)
 Breaking changes: