+++ CHANGELOG.md
@@ -1,8 +1,11 @@
 # Unreleased
-- [timePicker] Fix display on small devices
-- Remove deprecated updateColor() method
-- [itemChoice] Fix right content display
+- [...]
+
+# v0.3.1 (20/06/2018)
+- [TimePicker] Fix display on small devices
+- [ItemChoice] Fix right content display
 - [Checkbox] Reverse label and input and made it configurable
+- Remove deprecated updateColor() method
 
 # v0.3.0 (08/06/2018)
 Breaking changes: